Refactor kudu load generator functionality
