+++ src/main/scala/io/phdata/retirementage/loadgen/LoadGenerator.scala
@@ -18,10 +18,15 @@ package io.phdata.retirementage.loadgen
 
 import java.util.UUID
 
+import io.phdata.retirementage.domain.GlobalConfig
+import org.apache.kudu.client.CreateTableOptions
+import org.apache.kudu.spark.kudu.KuduContext
 import org.apache.spark.sql.types.{StringType, StructField, StructType}
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions.lit
 import org.apache.spark.storage.StorageLevel
+
+import scala.collection.JavaConverters._
 
 object LoadGenerator {
   def main(args: Array[String]): Unit = {
@@ -37,29 +42,24 @@ object LoadGenerator {
       .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
       .getOrCreate()
 
-    generateTables(spark, conf)
+    GlobalConfig.kuduMasters = Some(List("master3.valhalla.phdata.io:7051"))
+
+    val kuduContext =
+      new KuduContext(GlobalConfig.kuduMasters.get.mkString(","), spark.sqlContext.sparkContext)
+
+    generateTables(spark, kuduContext, conf)
   }
 
-  def generateTables(spark: SparkSession, conf: LoadGeneratorConfig): Unit = {
+  def generateTables(spark: SparkSession,
+                     kuduContext: KuduContext,
+                     conf: LoadGeneratorConfig): Unit = {
     // generate fact table
-    val factdf = generateTable(spark, conf.factCount(), 1).persist(StorageLevel.MEMORY_AND_DISK)
-
-    factdf
-      .coalesce(50)
-      .write
-      .mode(SaveMode.Overwrite)
-      .saveAsTable(s"${conf.databaseName()}.${conf.factName()}")
+    val factDf = generateTable(spark, conf.factCount(), 1).persist(StorageLevel.MEMORY_AND_DISK)
 
     // generate dimension table
     val dimensionDf =
-      generateTableFromParent(spark, conf.dimensionCount(), 2, conf.factCount(), factdf)
+      generateTableFromParent(spark, conf.dimensionCount(), 2, conf.factCount(), factDf)
         .persist(StorageLevel.MEMORY_AND_DISK)
-
-    dimensionDf
-      .repartition(25)
-      .write
-      .mode(SaveMode.Overwrite)
-      .saveAsTable(s"${conf.databaseName()}.${conf.dimName()}")
 
     // generate subdimension table
     val subDimensionDf = generateTableFromParent(spark,
@@ -68,11 +68,37 @@ object LoadGenerator {
                                                  conf.dimensionCount(),
                                                  dimensionDf).persist(StorageLevel.MEMORY_AND_DISK)
 
-    subDimensionDf
-      .repartition(25)
-      .write
-      .mode(SaveMode.Overwrite)
-      .saveAsTable(s"${conf.databaseName()}.${conf.subName()}")
+    conf.outputFormat() match {
+      case "parquet" =>
+        factDf
+          .coalesce(50)
+          .write
+          .mode(SaveMode.Overwrite)
+          .saveAsTable(s"${conf.databaseName()}.${conf.factName()}")
+        dimensionDf
+          .repartition(25)
+          .write
+          .mode(SaveMode.Overwrite)
+          .saveAsTable(s"${conf.databaseName()}.${conf.dimName()}")
+        subDimensionDf
+          .repartition(25)
+          .write
+          .mode(SaveMode.Overwrite)
+          .saveAsTable(s"${conf.databaseName()}.${conf.subName()}")
+      case "kudu" =>
+        createKuduTable(kuduContext, conf.factName(), 10)
+
+        kuduContext.insertRows(factDf, conf.factName())
+
+        createKuduTable(kuduContext, conf.dimName(), 10)
+
+        kuduContext.insertRows(dimensionDf, conf.dimName())
+
+        createKuduTable(kuduContext, conf.subName(), 10)
+
+        kuduContext.insertRows(subDimensionDf, conf.subName())
+      case _ => throw new UnsupportedOperationException
+    }
   }
 
   case class LoadGenTemp(id: String, payload: String, dimension_id: String, expiration_date: String)
@@ -192,5 +218,27 @@ object LoadGenerator {
       finalDf
     }
 
+  }
+
+  /*
+   * createKuduTable creates a kudu table given:
+   * tableName: The name of the kudu table to create
+   * numBuckets: The number of buckets for the hash partition
+   */
+  def createKuduTable(kuduContext: KuduContext, tableName: String, numBuckets: Int): Unit = {
+    val defaultSchema = StructType(StructField("id", StringType, false) :: Nil)
+      .add(StructField("payload", StringType, false))
+      .add(StructField("dimension_id", StringType, false))
+      .add(StructField("expiration_date", StringType, false))
+    val defaultKey = Seq("id")
+
+    val tableOptions = new CreateTableOptions()
+      .setNumReplicas(1)
+      .addHashPartitions(List(defaultKey(0)).asJava, numBuckets)
+
+    if (kuduContext.tableExists(tableName)) {
+      kuduContext.deleteTable(tableName)
+    }
+    kuduContext.createTable(tableName, defaultSchema, defaultKey, tableOptions)
   }
 }
